# Vakaros.Vkx.Api/OpenApi/Vakaros.Vkx.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GetDocument.Insider | v1",
    "description": "GetDocument Command-line Tool inside man",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/races/{raceId}": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDetailDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRaceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRaceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/positions": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/wind": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/speed-through-water": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/depth": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/temperature": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/load": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry/shift-angles": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/races/{raceId}/telemetry": {
      "get": {
        "tags": [
          "Races"
        ],
        "summary": "Fetches all telemetry channels in a single request.",
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RaceTelemetryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceTelemetryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceTelemetryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/races/{raceId}/analysis/start-line-length": {
      "get": {
        "tags": [
          "Races"
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StartLineLengthDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartLineLengthDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartLineLengthDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LinePositionDto": {
        "required": [
          "time",
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "PatchRaceRequest": {
        "required": [
          "courseId",
          "notes"
        ],
        "type": "object",
        "properties": {
          "courseId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RaceDetailDto": {
        "required": [
          "id",
          "sessionId",
          "raceNumber",
          "courseId",
          "countdownStartedAt",
          "countdownDurationSeconds",
          "startedAt",
          "endedAt",
          "durationSeconds",
          "sailedDistanceMeters",
          "maxSpeedOverGround",
          "notes",
          "pinEnd",
          "boatEnd",
          "startAnalysis"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "raceNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "courseId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "countdownStartedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "countdownDurationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "durationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sailedDistanceMeters": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxSpeedOverGround": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "float"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "pinEnd": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LinePositionDto"
              }
            ]
          },
          "boatEnd": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LinePositionDto"
              }
            ]
          },
          "startAnalysis": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StartAnalysisDto"
              }
            ]
          }
        }
      },
      "RaceDto": {
        "required": [
          "id",
          "raceNumber",
          "courseId",
          "courseName",
          "countdownStartedAt",
          "countdownDurationSeconds",
          "startedAt",
          "endedAt",
          "durationSeconds",
          "sailedDistanceMeters",
          "maxSpeedOverGround",
          "notes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "raceNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "courseId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "courseName": {
            "type": [
              "null",
              "string"
            ]
          },
          "countdownStartedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "countdownDurationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "durationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sailedDistanceMeters": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxSpeedOverGround": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "float"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RaceTelemetryDto": {
        "required": [
          "positions",
          "wind",
          "speedThroughWater",
          "depth",
          "temperature",
          "load",
          "shiftAngles"
        ],
        "type": "object",
        "properties": {
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionDto"
            }
          },
          "wind": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WindDto"
            }
          },
          "speedThroughWater": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpeedThroughWaterDto"
            }
          },
          "depth": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepthDto"
            }
          },
          "temperature": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemperatureDto"
            }
          },
          "load": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadDto"
            }
          },
          "shiftAngles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftAngleDto"
            }
          }
        }
      },
...
      "StartLineLengthDto": {
        "required": [
          "lengthMeters"
        ],
        "type": "object",
        "properties": {
          "lengthMeters": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
    }
  },
...
}